move files into client + cache packages
