+++ src/client/client.go
@@ -1,4 +1,4 @@
-package cache
+package client
 
 /************************************************
 Client supports
